requirements.txt
streamlit
openai
tavily-python
gTTS
elevenlabs